Update li21a paper and supplementary material.
My name is Yuantong Li and I am one of the authors of the paper “Online Forgetting Process for Linear Regression Models”.  We have fixed “norm” symbol and the missing lemma and theorem symbol in appendix for the camera-ready version. It would be great if you could update the current version with these changes.

Thank you very much for your kind help! Have a great day!

Best,
Yuantong Li